// src/payload-types.ts
export interface Config {
  auth: {
    users: UserAuthOperations;
  };
  blocks: {
    content: ContentBlock;
    mediaBlock: MediaBlock;
    archive: ArchiveBlock;
    formBlock: FormBlock;
    hero_one: HeroOneBlock;
    hero_two: HeroTwoBlock;
    hero_three: HeroThreeBlock;
    hero_four: HeroFourBlock;
    hero_five: HeroFiveBlock;
    about_us: AboutUsBlock;
    about_us_two: AboutUsTwoBlock;
    about_us_three: AboutUsThreeBlock;
    about_us_four: AboutUsFourBlock;
    counters: CountersBlock;
    counters_two: CountersTwoBlock;
    services: ServicesBlock;
    services_two: ServicesTwoBlock;
    services_three: ServicesThreeBlock;
    services_four: ServicesFourBlock;
    call_to_action: CallToActionBlock;
    team_block: TeamBlock;
    call_to_action_2: CallToAction2Block;
    subscribe: SubscribeBlock;
    portfolio: PortfolioBlock;
    portfolio_two: PortfolioTwoBlock;
    prices_block: PricesBlock;
    prices_two: PricesTwoBlock;
    intro: IntroBlock;
    portfolio_grid: PortfolioGrid;
    about_project: AboutProjectBlock;
    about_service: AboutServiceBlock;
    features: FeaturesBlock;
    features_two: FeaturesTwoBlock;
    faq: FaqBlock;
    reusable_block: ReusableBlock;
    about_team: AboutTeamBlock;
    testimonials: TestimonialsBlock;
    steps_block: StepsBlock;
    contact_form_block: ContactFormBlock;
    clients_block: ClientsBlock;
    career_block: CareerBlock;
    support_grid_block: SupportGridBlock;
    commitment_block: CommitmentBlock;
    staffing_services_block: StaffingServicesBlock;
    it_services_block: ITServicesBlock;
    industries_block: IndustriesBlock;
    why_partner_block: WhyPartnerBlock;
    about_us_premium_block: AboutUsPremiumBlock;
    core_services_block: CoreServicesBlock;
    premium_overview_block: PremiumOverviewBlock;
  };
  collections: {
    pages: Page;
    posts: Post;
    projects: Project;
    services: Service;
    team: Team;
    media: Media;
    categories: Category;
    users: User;
    'reusable-content': ReusableContent;
    clients: Client;
    redirects: Redirect;
    forms: Form;
    'form-submissions': FormSubmission;
    'payload-kv': PayloadKv;
    'payload-jobs': PayloadJob;
    'payload-folders': FolderInterface;
    'payload-locked-documents': PayloadLockedDocument;
    'payload-preferences': PayloadPreference;
    'payload-migrations': PayloadMigration;
  };
  collectionsJoins: {
    'payload-folders': {
      documentsAndFolders: 'payload-folders' | 'media';
    };
  };
  collectionsSelect: {
    pages: PagesSelect<false> | PagesSelect<true>;
    posts: PostsSelect<false> | PostsSelect<true>;
    projects: ProjectsSelect<false> | ProjectsSelect<true>;
    services: ServicesSelect<false> | ServicesSelect<true>;
    team: TeamSelect<false> | TeamSelect<true>;
    media: MediaSelect<false> | MediaSelect<true>;
    categories: CategoriesSelect<false> | CategoriesSelect<true>;
    users: UsersSelect<false> | UsersSelect<true>;
    'reusable-content': ReusableContentSelect<false> | ReusableContentSelect<true>;
    clients: ClientsSelect<false> | ClientsSelect<true>;
    redirects: RedirectsSelect<false> | RedirectsSelect<true>;
    forms: FormsSelect<false> | FormsSelect<true>;
    'form-submissions': FormSubmissionsSelect<false> | FormSubmissionsSelect<true>;
    'payload-kv': PayloadKvSelect<false> | PayloadKvSelect<true>;
    'payload-jobs': PayloadJobsSelect<false> | PayloadJobsSelect<true>;
    'payload-folders': PayloadFoldersSelect<false> | PayloadFoldersSelect<true>;
    'payload-locked-documents': PayloadLockedDocumentsSelect<false> | PayloadLockedDocumentsSelect<true>;
    'payload-preferences': PayloadPreferencesSelect<false> | PayloadPreferencesSelect<true>;
    'payload-migrations': PayloadMigrationsSelect<false> | PayloadMigrationsSelect<true>;
  };
  db: {
    defaultIDType: string;
  };
  fallbackLocale: null;
  globals: {
    header: Header;
    footer: Footer;
    blog: Blog;
    services_page: ServicesPage;
    projects_page: ProjectsPage;
    team_page: TeamPage;
    general: General;
  };
  globalsSelect: {
    header: HeaderSelect<false> | HeaderSelect<true>;
    footer: FooterSelect<false> | FooterSelect<true>;
    blog: BlogSelect<false> | BlogSelect<true>;
    services_page: ServicesPageSelect<false> | ServicesPageSelect<true>;
    projects_page: ProjectsPageSelect<false> | ProjectsPageSelect<true>;
    team_page: TeamPageSelect<false> | TeamPageSelect<true>;
    general: GeneralSelect<false> | GeneralSelect<true>;
  };
  locale: null;
  user: User & {
    collection: 'users';
  };
  jobs: {
    tasks: {
      schedulePublish: TaskSchedulePublish;
      inline: {
        input: unknown;
        output: unknown;
      };
    };
    workflows: unknown;
  };
}
export interface UserAuthOperations {
  forgotPassword: {
    email: string;
    password: string;
  };
  login: {
    email: string;
    password: string;
  };
  registerFirstUser: {
    email: string;
    password: string;
  };
  unlock: {
    email: string;
    password: string;
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContentBlock".
 */
export interface ContentBlock {
  title?: string | null;
  columns?:
    | {
        size?: ('oneThird' | 'half' | 'twoThirds' | 'full') | null;
        richText?: {
          root: {
            type: string;
            children: {
              type: any;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        } | null;
        id?: string | null;
      }[]
    | null;
  dividerTop?: boolean | null;
  dividerBottom?: boolean | null;
  padding_top?: ('0' | '160') | null;
  padding_bottom?: ('0' | '40' | '100' | '130' | '145' | '160') | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'content';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "MediaBlock".
 */
export interface MediaBlock {
  media: string | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'mediaBlock';
}
// ...
export interface ArchiveBlock {
  badge?: string | null;
  title?: string | null;
  description?: string | null;
  button_text?: string | null;
  button_url?: string | null;
  populateBy?: ('collection' | 'selection') | null;
  relationTo?: 'posts' | null;
  categories?: (string | Category)[] | null;
  limit?: number | null;
  selectedDocs?:
    | {
        relationTo: 'posts';
        value: string | Post;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'archive';
}
// ...
export interface FormBlock {
  badge?: string | null;
  title?: string | null;
  contact_items?:
    | {
        label?: string | null;
        value?: string | null;
        id?: string | null;
      }[]
    | null;
  social_links?:
    | {
        icon_class?: string | null;
        url?: string | null;
        id?: string | null;
      }[]
    | null;
  desc_title?: string | null;
  desc_content?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  form_title?: string | null;
  form: string | Form;
  gmap_title?: string | null;
  gmap?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'formBlock';
}
// ...
export interface HeroOneBlock {
  bgImage?: (string | null) | Media;
  icon?: (string | null) | Media;
  subtitle?: string | null;
  title_line1?: string | null;
  title_line2?: string | null;
  title_line3?: string | null;
  description?: string | null;
  button1_text?: string | null;
  button1_url?: string | null;
  button2_text?: string | null;
  button2_url?: string | null;
  counter1_upper?: string | null;
  counter1_lower?: string | null;
  counter1_value?: string | null;
  counter2_upper?: string | null;
  counter2_lower?: string | null;
  counter2_value?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'hero_one';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeroTwoBlock".
 */
export interface HeroTwoBlock {
  title?: string | null;
  users?:
    | {
        image: string | Media;
        id?: string | null;
      }[]
    | null;
  description?: string | null;
  button_text?: string | null;
  button_url?: string | null;
  bgImage?: (string | null) | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'hero_two';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeroThreeBlock".
 */
export interface HeroThreeBlock {
  video?: (string | null) | Media;
  title?: string | null;
  button1_text?: string | null;
  button1_url?: string | null;
  button2_text?: string | null;
  button2_url?: string | null;
  description?: string | null;
  users?:
    | {
        image: string | Media;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'hero_three';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeroFourBlock".
 */
export interface HeroFourBlock {
  icon?: (string | null) | Media;
  subtitle?: string | null;
  title?: string | null;
  description?: string | null;
  button1_text?: string | null;
  button1_url?: string | null;
  button2_text?: string | null;
  button2_url?: string | null;
  bgImage?: (string | null) | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'hero_four';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeroFiveBlock".
 */
export interface HeroFiveBlock {
  icon?: (string | null) | Media;
  subtitle?: string | null;
  slides?:
    | {
        bgImage: string | Media;
        title: string;
        id?: string | null;
      }[]
    | null;
  description?: string | null;
  button1_text?: string | null;
  button1_url?: string | null;
  button2_text?: string | null;
  button2_url?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'hero_five';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "AboutUsBlock".
 */
export interface AboutUsBlock {
  badge?: string | null;
  title?: string | null;
  description?: string | null;
  portrait?: (string | null) | Media;
  founder_name?: string | null;
  founder_role?: string | null;
  founder_quote?: string | null;
  youtube_video_link?: string | null;
  image?: (string | null) | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'about_us';
}
// ...
export interface CountersBlock {
  counters?:
    | {
        value: string;
        suffix?: string | null;
        /**
         * Optional CSS class for suffix styling
         */
        suffix_class?: string | null;
        title: string;
        id?: string | null;
      }[]
    | null;
  description?: string | null;
  image?: (string | null) | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'counters';
}
// ...
export interface ReusableContent {
  id: string;
  title: string;
  layout: (
    | ContentBlock
    | MediaBlock
    | ArchiveBlock
    | FormBlock
    | HeroOneBlock
    | HeroTwoBlock
    | HeroThreeBlock
    | HeroFourBlock
    | HeroFiveBlock
    | AboutUsBlock
    | AboutUsTwoBlock
    | AboutUsThreeBlock
    | AboutUsFourBlock
    | CountersBlock
    | CountersTwoBlock
    | ServicesBlock
    | ServicesTwoBlock
    | ServicesThreeBlock
    | ServicesFourBlock
    | CallToActionBlock
    | TeamBlock
    | CallToAction2Block
    | SubscribeBlock
    | PortfolioBlock
    | PortfolioTwoBlock
    | PricesBlock
    | PricesTwoBlock
    | IntroBlock
    | PortfolioGrid
    | AboutProjectBlock
    | AboutServiceBlock
    | AboutTeamBlock
    | FeaturesBlock
    | FeaturesTwoBlock
    | FaqBlock
    | TestimonialsBlock
    | StepsBlock
    | ContactFormBlock
    | ClientsBlock
    | CareerBlock
    | SupportGridBlock
    | CommitmentBlock
    | StaffingServicesBlock
    | ITServicesBlock
    | IndustriesBlock
    | WhyPartnerBlock
    | AboutUsPremiumBlock
    | CoreServicesBlock
    | PremiumOverviewBlock
  )[];
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
// ...
export interface ClientsSelect<T extends boolean = true> {
  name?: T;
  logo?: T;
  website?: T;
  updatedAt?: T;
  createdAt?: T;
}